+++ src/Dfe.PlanTech.Application/Questionnaire/Queries/GetNextUnansweredQuestionQuery.cs
@@ -1,7 +1,7 @@
 using Dfe.PlanTech.Application.Exceptions;
 using Dfe.PlanTech.Domain.Questionnaire.Interfaces;
 using Dfe.PlanTech.Domain.Questionnaire.Models;
-using Dfe.PlanTech.Domain.Responses.Interface;
+using Dfe.PlanTech.Domain.Responses.Interfaces;
 
 namespace Dfe.PlanTech.Application.Questionnaire.Queries;
