+++ src/Dfe.PlanTech.Application/Responses/Commands/ProcessCheckAnswerDtoCommand.cs
@@ -1,6 +1,5 @@
 using Dfe.PlanTech.Application.Exceptions;
 using Dfe.PlanTech.Domain.Questionnaire.Models;
-using Dfe.PlanTech.Domain.Responses.Interface;
 using Dfe.PlanTech.Domain.Responses.Interfaces;
 
 namespace Dfe.PlanTech.Application.Responses.Commands;
