+++ src/Dfe.PlanTech.Application/Responses/Queries/GetLatestResponsesQuery.cs
@@ -1,6 +1,6 @@
 using Dfe.PlanTech.Application.Persistence.Interfaces;
 using Dfe.PlanTech.Domain.Questionnaire.Models;
-using Dfe.PlanTech.Domain.Responses.Interface;
+using Dfe.PlanTech.Domain.Responses.Interfaces;
 using Dfe.PlanTech.Domain.Submissions.Models;
 using System.Linq.Expressions;
