+++ src/Dfe.PlanTech.Web/Controllers/CheckAnswersController.cs
@@ -1,5 +1,3 @@
-using System.Data.Common;
-using Dfe.PlanTech.Application.Constants;
 using Dfe.PlanTech.Application.Content.Queries;
 using Dfe.PlanTech.Application.Exceptions;
 using Dfe.PlanTech.Domain.Content.Models;
